+++ server/component/style/qualtricsSurvey/QualtricsSurveyView.php
@@ -100,14 +100,13 @@ class QualtricsSurveyView extends StyleView
     {
         $style = parent::output_content_mobile();
         $style['qualtrics_url'] = $this->model->get_survey_link();
+        $style['alert'] = '';
         if ($this->model->is_survey_active()) {
             if ($this->model->is_survey_done()) {
                 $style['alert'] = $this->label_survey_done;
             }
         } else {
-            if ($this->label_survey_not_active != '') {
-                $style['alert'] = $this->label_survey_not_active;
-            }
+            $style['alert'] = $this->label_survey_not_active;
         }
         return $style;
     }